Fix view annotation management.
Removing views hasn't been working previously as we werent deleting annotation from the map. Also adding new annotation in the real time was buggy as we should've add the new annotation to views registry first before adding it to the map. Otherwise when added to map viewForAnnotation gets triggered and it's unable to look up the view in which case the default view hes been redered.
Last issue was the problem with some of the map delegate methods that were assuming RCTMGLAnnotation is the only possible annotation type whereas RCTMapboxAnnotation could also be provided.